pyproject.toml
[tool.poetry]
name = "testcontainers_demo_generator"
version = "0.1.0"
description = "A tool that generates Testcontainers demo applications"
authors = ["Anna Chernyshova"]
packages = [
    { include = "testcontainers_demo_generator", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
crewai = "*"
anthropic = "*"
python-dotenv = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

